We need to keep the media recorder instance as a Ref
Using it as part of the rendering state causes a lot of re-mounting of
the component because of dependencies of useEffect hooks.

The downside of putting the recorder behind a ref is that we need a
separate piece of state (`recording` in this case`) for rendering
purposes, since changes to refs' contents do not trigger re-renders.